Fix autofocus bugs
