+++ app/javascript/components/CardColumn/Card/card-popup/card-popup.css
@@ -8,6 +8,7 @@
   bottom: 0;
   margin: auto;
   background-color: rgba(0,0,0, 0.5);
+  z-index: 100;
 }
 
 .card-popup__inner {
