+++ app/views/boards/show.html.erb
@@ -43,7 +43,7 @@
   <% column_class_name = 'column is-one-quarter' %>
 <% end %>
 
-<%= react_component('card_table', {
+<%= react_component('card-table', {
   actionItems: @action_items,
   board: @board,
   cardsByType: @cards_by_type,
