+++ package.json
@@ -79,7 +79,7 @@
     "react-bulma-components": "3.1.3",
     "react-dom": "^16.12.0",
     "react-select": "^3.0.8",
-    "react-textarea-autosize": "^7.1.2",
+    "react-textarea-autosize": "8.3.0",
     "react_ujs": "^2.6.1"
   },
   "devDependencies": {
